Extract OAuth2 callback in separate file
