+++ src/index.js
@@ -5,24 +5,10 @@ import { tokenRepository } from './tokenRepository.js';
 import Accounts from './components/Accounts.jsx';
 import TokenStore from './components/TokenStore.jsx';
 import platforms from './platforms';
-
-const platformOauth2Callback = function (uri) {
-    const request = JSON.parse(localStorage.getItem('oauth2.request'));
-    const platform = platforms.filter(p => p.id === request.platform)[0];
-    platform.client.token.getToken(uri, {state: request.state})
-        .then(function (user) {
-            console.log(user);
-            tokenRepository.addToken(platform, user);
-            location.href = '/';
-        })
-        .catch(function (err) {
-            console.log(err);
-        })
-    ;
-}
+import platformOAuth2Callback from './platformOAuth2Callback';
 
 if (location.pathname === '/redirect') {
-    platformOauth2Callback(location.href);
+    platformOAuth2Callback(location.href);
 }
 
 render(
